reorganized folder structure and temporarily disabled reaction logic
