+++ src/main/java/com/github/waifu/chat/commands/SimpleHandler.java
@@ -1,4 +1,4 @@
-package com.github.waifu.commands;
+package com.github.waifu.chat.commands;
 
 import com.google.gson.JsonObject;
