+++ src/main/java/com/github/waifu/chat/commands/TenorHandler.java
@@ -1,4 +1,4 @@
-package com.github.waifu.commands;
+package com.github.waifu.chat.commands;
 
 import org.json.JSONObject;
